+++ libs/react/src/__mocks__/customHandlers.ts
@@ -228,4 +228,45 @@ export const getActionHandler = {
       } as ActionPayload),
     );
   }),
+  subcategoryActionsOverviewCard: [
+    rest.get('*/organizations/*/actions', (req, res, ctx) => {
+      // return actions with some of them having the ready_to_execute and all survey submitted
+      const facilitiesActions = getActionsMock().filter(
+        (action) => action.action.subcategory === 'facilities',
+      );
+      facilitiesActions[0].action.dependent_surveys.forEach((survey, index) => {
+        survey.submitted = true;
+      });
+      facilitiesActions[1].action.dependent_surveys.forEach((survey) => {
+        survey.submitted = true;
+      });
+      facilitiesActions[0].action.ready_to_execute = true;
+      facilitiesActions[1].action.ready_to_execute = false;
+      const actions = getActionsMock().map((actionPayload, index) => {
+        if (actionPayload.action.subcategory === 'facilities') {
+          if (facilitiesActions[0].id === actionPayload.id) {
+            return facilitiesActions[0];
+          } else {
+            return facilitiesActions[1];
+          }
+        } else {
+          return actionPayload;
+        }
+      });
+      return res(ctx.json(actions));
+    }),
+  ],
+  actionsOverview: [
+    rest.get('*/organizations/*/actions', (req, res, ctx) => {
+      const actions = getActionsMock().map((actionPayload, index) => {
+        // set all the surveys to submitted and ready to execute to true
+        actionPayload.action.dependent_surveys.forEach((survey) => {
+          survey.submitted = true;
+        });
+        actionPayload.action.ready_to_execute = true;
+        return actionPayload;
+      });
+      return res(ctx.json(actions));
+    }),
+  ],
 };